parse/requirements.txt
GitPython==3.1.20
pandas==0.25.1
